requirements.txt
apturl==0.5.2
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.8
cupshelpers==1.0
cvzone==1.6.1
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.8
distro==1.4.0
distro-info==0.23+ubuntu1.1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.15.4
future==0.18.2
graphsurgeon==0.4.6
httplib2==0.14.0
idna==2.8
jetson-stats==4.2.8
Jetson.GPIO==2.1.1
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.1.2
monotonic==1.5
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
onnx-graphsurgeon==0.3.12
opencv-python==4.10.0.84
pandas==0.25.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==4.2.2
protobuf==3.6.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-dbusmock==0.19
python-debian==0.1.36+ubuntu1.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
scipy==1.3.3
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
smbus2==0.4.3
systemd-python==234
tensorrt==8.5.2.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uff==0.6.9
urllib3==1.25.8
urwid==2.0.1
virtualenv==20.26.3
wadllib==1.3.3
xkit==0.0.0
